+++ lib/esse/rails.rb
@@ -2,6 +2,8 @@
 
 require "esse"
 require "active_support"
+require "rails/railtie"
+require "esse/rails/instrumentation"
 
 module Esse
   module Rails